:necktie: Use real fetch api
