+++ src/app/api/notes/[id]/route.ts
@@ -1,29 +1,10 @@
 import { NextResponse } from 'next/server'
-import { NoteData } from '../../../server-actions-examples/_lib/actions'
-
-// メモリ内のノートデータストア（実際のアプリではデータベースを使用）
-// 注意: 実際のアプリでは、このデータは親ルートと共有する必要があります
-let notes: NoteData[] = [
-  {
-    content: 'App Routerは、Next.js 13から導入された新しいルーティングシステムです。',
-    createdAt: new Date(Date.now() - 86400000).toISOString(),
-    id: '1',
-    title: 'Next.js App Routerについて', // 1日前
-    updatedAt: new Date(Date.now() - 86400000).toISOString(),
-  },
-  {
-    content:
-      'Server Actionsは、クライアントコンポーネントからサーバー関数を直接呼び出せる機能です。',
-    createdAt: new Date(Date.now() - 43200000).toISOString(),
-    id: '2',
-    title: 'Server Actionsの基本', // 12時間前
-    updatedAt: new Date(Date.now() - 43200000).toISOString(),
-  },
-]
+import { notes } from 'constants/note'
+import { Note } from 'core/domain/entity/note'
 
 // GET /api/notes/[id] - 特定のノートを取得
 export async function GET(request: Request, { params }: { params: { id: string } }) {
-  const id = params.id
+  const { id } = await params
   const note = notes.find((note) => note.id === id)
 
   if (!note) {
@@ -36,7 +17,7 @@ export async function GET(request: Request, { params }: { params: { id: string }
 // PUT /api/notes/[id] - 特定のノートを更新
 export async function PUT(request: Request, { params }: { params: { id: string } }) {
   try {
-    const id = params.id
+    const { id } = await params
     const body = await request.json()
 
     // 入力検証
@@ -51,15 +32,12 @@ export async function PUT(request: Request, { params }: { params: { id: string }
     }
 
     // ノートを更新
-    const updatedNote: NoteData = {
+    const updatedNote: Note = {
       ...notes[noteIndex],
       content: body.content || notes[noteIndex].content,
       title: body.title,
       updatedAt: new Date().toISOString(),
     }
-
-    // 更新したノートを保存
-    notes = [...notes.slice(0, noteIndex), updatedNote, ...notes.slice(noteIndex + 1)]
 
     return NextResponse.json(updatedNote)
   } catch (error) {
@@ -71,16 +49,13 @@ export async function PUT(request: Request, { params }: { params: { id: string }
 // DELETE /api/notes/[id] - 特定のノートを削除
 export async function DELETE(request: Request, { params }: { params: { id: string } }) {
   try {
-    const id = params.id
+    const { id } = await params
 
     // ノートを検索
     const noteIndex = notes.findIndex((note) => note.id === id)
     if (noteIndex === -1) {
       return NextResponse.json({ error: 'ノートが見つかりません' }, { status: 404 })
     }
-
-    // ノートを削除
-    notes = [...notes.slice(0, noteIndex), ...notes.slice(noteIndex + 1)]
 
     return NextResponse.json({ success: true })
   } catch (error) {
