+++ src/core/domain/entity/note.ts
@@ -1,0 +1,7 @@
+export interface Note {
+  id: string
+  title: string
+  content: string
+  createdAt: string
+  updatedAt: string
+}
